feat: add FormCard component
