+++ src/components/index.ts
@@ -1,4 +1,7 @@
 export { Header } from './Header/Header';
+export { Nav } from './Nav/Nav';
 export { Search } from './Search/Search';
 export { ProductCard } from './ProductCard/ProductCard';
 export { Form } from './Form/Form';
+export { FormCard } from './FormCard/FormCard';
+export { Message } from './Message/Message';
